+++ crates/sqlexec/src/context.rs
@@ -9,9 +9,9 @@ use crate::parser::{CustomParser, StatementWithExtensions};
 use crate::planner::errors::PlanError;
 use crate::planner::logical_plan::*;
 use crate::planner::session_planner::SessionPlanner;
-use crate::remote::broadcast::staged_stream::StagedClientStreams;
 use crate::remote::client::RemoteSessionClient;
 use crate::remote::planner::RemoteLogicalPlanner;
+use crate::remote::staged_stream::StagedClientStreams;
 use datafusion::arrow::datatypes::{DataType, Field as ArrowField, Schema as ArrowSchema};
 use datafusion::arrow::record_batch::RecordBatch;
 use datafusion::common::{Column as DfColumn, SchemaReference};
