+++ crates/sqlexec/src/planner/mod.rs
@@ -2,6 +2,7 @@ pub mod dispatch;
 pub mod errors;
 pub mod extension;
 pub mod logical_plan;
+pub mod physical_plan;
 pub mod session_planner;
 
 pub(crate) mod context_builder;
