--- crates/sqlexec/src/remote/broadcast/mod.rs
@@ -1,3 +1,0 @@
-//! Broadcast implementation (moving batches between client and server).
-pub mod exchange_exec;
-pub mod staged_stream;
