+++ crates/sqlexec/src/remote/mod.rs
@@ -1,5 +1,6 @@
-pub mod broadcast;
 pub mod client;
+pub mod exchange_stream;
 pub mod exec;
 pub mod planner;
+pub mod staged_stream;
 pub mod table;
